+++ Sources/WalletConnectVerify/VerifyClientFactory.swift
@@ -17,10 +17,15 @@ public class VerifyClientFactory {
             attestChallengeProvider: attestChallengeProvider,
             keyAttestationService: keyAttestationService
         )
+        let verifyServerPubKeyManagerStore: CodableStore<PublicKeyFetcher.VerifyServerPublicKey> = CodableStore(defaults: keyValueStorage, identifier: "com.walletconnect.verify")
+
+        let verifyServerPubKeyManager = VerifyServerPubKeyManager(store: verifyServerPubKeyManagerStore)
+        let attestationVerifier = AttestationJWTVerifier(verifyServerPubKeyManager: verifyServerPubKeyManager)
         return VerifyClient(
             originVerifier: originVerifier,
             assertionRegistrer: assertionRegistrer,
-            appAttestationRegistrer: appAttestationRegistrer
+            appAttestationRegistrer: appAttestationRegistrer,
+            attestationVerifier: attestationVerifier
         )
     }
 }
